requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.5.0
cryptography==2.4.2
enum34==1.1.6
flake8==3.6.0
Flask==1.0.2
Flask-Cors==3.0.7
futures==3.2.0
google-api-core==1.5.2
google-auth==1.6.1
google-cloud==0.34.0
google-cloud-core==0.28.1
google-cloud-datastore==1.7.1
google-cloud-logging==1.8.0
google-cloud-monitoring==0.31.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpcio==1.16.1
hashids==1.2.0
httplib2==0.12.0
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
mongoengine==0.16.3
ndb==1.0.13b1
ndg-httpsclient==0.5.1
oauth2client==1.5.0
proto-google-datastore-v1beta3==0.0.3rc0
protobuf==3.6.1
pusher==2.0.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.19
pycrypto==2.6.1
pyflakes==2.0.0
pymongo==3.7.2
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
requests-toolbelt==0.8.0
rsa==4.0
simplejson==3.16.0
six==1.11.0
testing-v1beta3-googledatastore==4.0.0b4
uritemplate==0.6
urllib3==1.24.1
Werkzeug==0.14.1
